+++ kikuchipy/data/__init__.py
@@ -32,6 +32,8 @@ desired.
 """
 
 __all__ = [
+    "ni_gain0",
+    "ni_gain0_calibration",
     "nickel_ebsd_large",
     "nickel_ebsd_master_pattern_small",
     "nickel_ebsd_small",
@@ -48,6 +50,8 @@ def __dir__():
 
 def __getattr__(name):
     _import_mapping = {
+        "ni_gain0": "_data",
+        "ni_gain0_calibration": "_data",
         "nickel_ebsd_large": "_data",
         "nickel_ebsd_master_pattern_small": "_data",
         "nickel_ebsd_small": "_data",